Introduce SystemsBase and HomotopiesBase to resolve dependency cycles
