+++ src/HomotopyContinuation.jl
@@ -22,7 +22,9 @@ module HomotopyContinuation
     include("parallel.jl")
     include("projective_vectors.jl")
     include("affine_patches.jl")
-    
+
+    include("systems_base.jl")
+    include("homotopies_base.jl")
     include("systems.jl")
     include("homotopies.jl")
     include("problems.jl")
